refactor(#62): Paging 관련 Class common 모듈로 변경
